unet 200 epochs random and stratified shuffling, trasnfer learning

+++ src/unet_settings.py
@@ -4,7 +4,8 @@ from pathlib import Path
 import pandas as pd
 
 ## SETTINGS
-config_name = 'unet256_randomshuffling'
+config_name = 'unet_256_l1/random_shuffling_pre_trained'
+
 Path(f'../{config_name}/models').mkdir(parents=True, exist_ok=True)
 Path(f'../{config_name}/figures').mkdir(exist_ok=True)
 
@@ -25,30 +26,33 @@ data_loading_settings = {
 
 model_settings = {
     'encoder_name': "efficientnet-b7",
-    'encoder_weights': None,
+    'encoder_weights': None, #"imagenet" or None
     'in_channels': 4,
     'classes': 6 if patch_level_param['level'] == 1 else 113, # 113 to be checked
     'path_to_intermed_model': f'../{config_name}/models/unet_intermed',
     'path_to_intermed_optim': f'../{config_name}/models/optim_intermed',
-    'path_to_model': f'../{config_name}/models/unet_last.pt',
-    'path_to_optim': f'../{config_name}/models/optim_last.pt',
+    'path_to_last_model': f'../{config_name}/models/unet_last.pt',
+    'path_to_last_optim': f'../{config_name}/models/optim_last.pt',
+    'path_to_best_model': f'../{config_name}/models/unet_intermed_epoch35.pt',#f'../{config_name}/models/unet_intermed_epoch3.pt',
 }
 
 training_settings = {
-    'training': True,
+    'training': False,
     'lr': 1e-4,
     'criterion': 'Dice', #Dice or CrossEntropy
     'optimizer': 'Adam',
     'nb_epochs': 200, 
-    'early_stopping': False,
-    'patience': 50,
-    'losss_path': f'../{config_name}/losses.csv ',
+    'early_stopping': True,
+    'patience': 30,
+    'losses_mious_path': f'../{config_name}/losses_mious.csv',
+    'restart_training': None, # 42 if you want to restart training from a certain epoch, put the epoch number here, else put 0
 }
 
 plotting_settings = {
     'plot_test': True,
     'pred_plot_path': f'../{config_name}/figures/test_preds.png',
     'losses_path': f'../{config_name}/figures/losses.png',
+    'mious_path': f'../{config_name}/figures/mious.png',
     'nb_plots': 6,
     'my_colors_map': {1: '#789262', 2: '#ff4500', 3: '#006400', 4: '#00ff00', 5: '#555555', 6: '#8a2be2'},
 }
@@ -75,8 +79,10 @@ settings = {
     'patience': training_settings['patience'],
 }
 
-# turn all to string
+# save settings to csv if not already done
+
 for key in settings.keys():
     settings[key] = str(settings[key])
 settings_df = pd.DataFrame(settings, index=[0])
-settings_df.to_csv(f'../{config_name}/settings.csv', index=False)
+if not Path(f'../{config_name}/settings.csv').exists():
+    settings_df.to_csv(f'../{config_name}/settings.csv', index=False)
